Refactor GitHub workflow for CMake multi-platform
This commit simplifies the GitHub actions workflow and renames it from 'CMake on multiple platforms' to 'C++ Builder'. The dropping of different compiler toolchains testing leads to a much cleaner and straight-forward workflow. Now it includes building on macOS, Windows and Ubuntu using default generator and compiler on each platform. The action also switches from version v3 to v2 for the checkout operation. Furthermore, the code is adjusted to make cmake version configurable and the workflow includes artifact archiving steps.